+++ src/core/core.mjs
@@ -1,5 +1,5 @@
 import path from 'path'
-import { applyColor, transformStackTrace } from '../utils/transform.mjs'
+import { transformStackTrace } from '../utils/transform.mjs'
 import { runParsedBlocks } from '../core/context.mjs'
 import { getTags, getReporterType, chooseTestFiles } from '../config/setup.mjs'
 import { timeStamp } from '../utils/support.mjs'
@@ -7,8 +7,9 @@ import { EXIT_CODES } from '../core/constants.mjs'
 import {
   printExecutionTime,
   printRunningTestFile,
-  printNewLine,
   printTags,
+  printFailuresMsg,
+  printTestResult,
 } from './output.mjs'
 import { getReporter } from '../reporters/index.mjs'
 
@@ -38,39 +39,4 @@ export const run = async () => {
     console.error(e.stack)
     process.exit(EXIT_CODES.failures)
   }
-}
-
-const createFullDescription = ({ name, describeStack }) =>
-  [...describeStack, { name }]
-    .map(({ name }) => `<bold>${name}</bold>`)
-    .join(' → ')
-
-const printFailureMsg = (failure) => {
-  console.error(applyColor(createFullDescription(failure)))
-  printNewLine()
-  failure.errors.forEach((error) => {
-    console.error(error.message)
-    console.error(error.stack)
-  })
-  printNewLine()
-}
-
-const printFailuresMsg = (failures) => {
-  if (failures.length > 0) {
-    printNewLine()
-    console.error('Failures:')
-    printNewLine()
-  }
-  failures.forEach(printFailureMsg)
-}
-
-const printTestResult = (failures, successes) => {
-  printNewLine()
-  console.log(
-    applyColor(
-      `Tests: <green>${successes} passed</green>, ` +
-        `<red>${failures.length} failed</red>, ` +
-        `${successes + failures.length} total`
-    )
-  )
 }
